requirements.txt
faiss-cpu==1.6.3
numpy==1.19.1
opencv-python==4.4.0.42
pillow==7.2.0
